splitted libtemp into 3 and fixed makefile
